// apps/api/src/auth/roles.enum.ts
export enum Role {
  ADMIN = 'ADMIN',
  LEADERSHIP = 'LEADERSHIP',
  OPS_MANAGER = 'OPS_MANAGER',
  BRANCH_MANAGER = 'BRANCH_MANAGER',
  RECOVERY_MANAGER = 'RECOVERY_MANAGER',
  CLAIMS_MANAGER = 'CLAIMS_MANAGER',
  SALES_MANAGER = 'SALES_MANAGER',
  FLEET_MANAGER = 'FLEET_MANAGER',
  FINANCE_MANAGER = 'FINANCE_MANAGER',
  CLAIMS_TEAM_IN = 'CLAIMS_TEAM_IN',
  CLAIMS_TEAM_OUT = 'CLAIMS_TEAM_OUT',
  CLAIMS_TEAM_LIABILITY = 'CLAIMS_TEAM_LIABILITY',
  CSE_DRIVER = 'CSE_DRIVER',
  SALES_REP = 'SALES_REP',
  RECOVERY_AGENT = 'RECOVERY_AGENT',
  FLEET_COORDINATOR = 'FLEET_COORDINATOR',
}